fix: move test to correct folder
