+++ tests/integration/Storefront/Framework/Captcha/CaptchaRouteListenerTest.php
@@ -2,31 +2,19 @@
 
 namespace Shopware\Tests\Integration\Storefront\Framework\Captcha;
 
-use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
-use Shopware\Core\Framework\Routing\KernelListenerPriorities;
 use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
 use Shopware\Core\Framework\Test\TestCaseBase\KernelLifecycleManager;
 use Shopware\Core\Framework\Test\TestCaseBase\SalesChannelApiTestBehaviour;
 use Shopware\Core\System\Salutation\SalutationCollection;
 use Shopware\Core\System\SystemConfig\SystemConfigService;
-use Shopware\Storefront\Framework\Captcha\AbstractCaptcha;
 use Shopware\Storefront\Framework\Captcha\BasicCaptcha;
-use Shopware\Storefront\Framework\Captcha\CaptchaException;
-use Shopware\Storefront\Framework\Captcha\CaptchaRouteListener;
 use Shopware\Storefront\Test\Controller\StorefrontControllerTestBehaviour;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\ParameterBag;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Event\ControllerEvent;
-use Symfony\Component\HttpKernel\HttpKernelInterface;
-use Symfony\Component\HttpKernel\KernelEvents;
-use Symfony\Component\Validator\ConstraintViolation;
-use Symfony\Component\Validator\ConstraintViolationList;
 
 /**
  * @internal
@@ -36,28 +24,6 @@ class CaptchaRouteListenerTest extends TestCase
     use IntegrationTestBehaviour;
     use SalesChannelApiTestBehaviour;
     use StorefrontControllerTestBehaviour;
-
-    public function testGetSubscribedEventsReturnsCorrectEvents(): void
-    {
-        static::assertSame([
-            KernelEvents::CONTROLLER => [
-                ['validateCaptcha', KernelListenerPriorities::KERNEL_CONTROLLER_EVENT_SCOPE_VALIDATE],
-            ],
-        ], CaptchaRouteListener::getSubscribedEvents());
-    }
-
-    public function testThrowsExceptionWhenValidationFails(): void
-    {
-        $event = $this->getControllerEventMock();
-
-        $this->expectExceptionMessage('The provided value for captcha');
-
-        (new CaptchaRouteListener(
-            $this->getCaptchas(true, false),
-            static::getContainer()->get(SystemConfigService::class),
-            static::getContainer()
-        ))->validateCaptcha($event);
-    }
 
     public function testJsonResponseWhenCaptchaValidationFails(): void
     {
@@ -141,240 +107,5 @@ class CaptchaRouteListenerTest extends TestCase
 
         static::assertInstanceOf(Response::class, $response);
         static::assertSame(200, $response->getStatusCode(), $response->getContent() ?: '');
-    }
-
-    public function testCaptchaSupportedButInvalidWithShouldBreakTrueAndNonXmlRequest(): void
-    {
-        $event = $this->getControllerEventMock();
-
-        $captcha = $this->getMockBuilder(AbstractCaptcha::class)->getMock();
-        $captcha->expects($this->once())
-            ->method('supports')
-            ->willReturn(true);
-        $captcha->expects($this->once())
-            ->method('isValid')
-            ->willReturn(false);
-        $captcha->expects($this->once())
-            ->method('shouldBreak')
-            ->willReturn(true);
-        $captcha->expects($this->once())
-            ->method('getViolations')
-            ->willReturn(new ConstraintViolationList());
-
-        $this->expectExceptionObject(CaptchaException::invalid($captcha));
-
-        (new CaptchaRouteListener(
-            [$captcha],
-            static::getContainer()->get(SystemConfigService::class),
-            static::getContainer()
-        ))->validateCaptcha($event);
-    }
-
-    public function testCaptchaSupportedButInvalidWithShouldBreakTrueAndXmlRequestWithNoViolations(): void
-    {
-        $request = new Request(
-            attributes: ['_captcha' => true],
-            server: ['HTTP_X_REQUESTED_WITH' => 'XMLHttpRequest']
-        );
-
-        $event = new ControllerEvent(
-            $this->createMock(HttpKernelInterface::class),
-            function (): void {},
-            $request,
-            HttpKernelInterface::MAIN_REQUEST
-        );
-
-        $captcha = $this->getMockBuilder(AbstractCaptcha::class)->getMock();
-        $captcha->expects($this->once())
-            ->method('supports')
-            ->willReturn(true);
-        $captcha->expects($this->once())
-            ->method('isValid')
-            ->willReturn(false);
-        $captcha->expects($this->once())
-            ->method('shouldBreak')
-            ->willReturn(true);
-
-        $violations = new ConstraintViolationList();
-        $captcha->expects($this->once())
-            ->method('getViolations')
-            ->willReturn($violations);
-
-        $listener = new CaptchaRouteListener(
-            [$captcha],
-            static::getContainer()->get(SystemConfigService::class),
-            static::getContainer()
-        );
-
-        $originalController = $event->getController();
-        $listener->validateCaptcha($event);
-
-        // Verify that a violation was added to the list with correct properties
-        // @see CaptchaRouteListener::validateCaptcha()
-        static::assertCount(1, $violations);
-        $violation = $violations->get(0);
-        static::assertInstanceOf(ConstraintViolation::class, $violation);
-
-        // Verify all properties set in the ConstraintViolation constructor
-        $expectedException = CaptchaException::invalid($captcha);
-        static::assertSame($expectedException->getMessage(), $violation->getMessage());
-        static::assertSame('Invalid captcha', $violation->getMessageTemplate());
-        static::assertSame($expectedException->getParameters(), $violation->getParameters());
-        static::assertSame('', $violation->getRoot());
-        static::assertSame('', $violation->getPropertyPath());
-        static::assertSame('', $violation->getInvalidValue());
-        static::assertNull($violation->getPlural());
-        static::assertSame($expectedException->getErrorCode(), $violation->getCode());
-
-        // Verify that the controller was changed to ErrorController
-        // @see CaptchaRouteListener::validateCaptcha()
-        static::assertNotSame($originalController, $event->getController());
-    }
-
-    public function testCaptchaSupportedButInvalidWithShouldBreakTrueAndXmlRequestWithExistingViolations(): void
-    {
-        $request = new Request(
-            attributes: ['_captcha' => true],
-            server: ['HTTP_X_REQUESTED_WITH' => 'XMLHttpRequest']
-        );
-
-        $event = new ControllerEvent(
-            $this->createMock(HttpKernelInterface::class),
-            function (): void {},
-            $request,
-            HttpKernelInterface::MAIN_REQUEST
-        );
-
-        $captcha = $this->getMockBuilder(AbstractCaptcha::class)->getMock();
-        $captcha->expects($this->once())
-            ->method('supports')
-            ->willReturn(true);
-        $captcha->expects($this->once())
-            ->method('isValid')
-            ->willReturn(false);
-        $captcha->expects($this->once())
-            ->method('shouldBreak')
-            ->willReturn(true);
-
-        $violations = new ConstraintViolationList();
-        $violations->add(new ConstraintViolation(
-            'Existing violation',
-            'Existing violation',
-            [],
-            '',
-            '',
-            ''
-        ));
-
-        $captcha->expects($this->once())
-            ->method('getViolations')
-            ->willReturn($violations);
-
-        $listener = new CaptchaRouteListener(
-            [$captcha],
-            static::getContainer()->get(SystemConfigService::class),
-            static::getContainer()
-        );
-
-        // Since violations count > 0, we expect an exception to be thrown
-        // @see CaptchaRouteListener::validateCaptcha()
-        $this->expectExceptionObject(CaptchaException::invalid($captcha));
-
-        $listener->validateCaptcha($event);
-    }
-
-    public function testCaptchaSupportedButInvalidWithShouldBreakFalseSetsErrorController(): void
-    {
-        $request = new Request(
-            query: ['_route' => 'frontend.home.page'],
-            attributes: ['_captcha' => true, '_route' => 'frontend.home.page']
-        );
-
-        $event = new ControllerEvent(
-            $this->createMock(HttpKernelInterface::class),
-            function (): void {},
-            $request,
-            HttpKernelInterface::MAIN_REQUEST
-        );
-
-        $captcha = $this->getMockBuilder(AbstractCaptcha::class)->getMock();
-        $captcha->expects($this->once())
-            ->method('supports')
-            ->willReturn(true);
-        $captcha->expects($this->once())
-            ->method('isValid')
-            ->willReturn(false);
-        $captcha->expects($this->once())
-            ->method('shouldBreak')
-            ->willReturn(false);
-
-        $violations = new ConstraintViolationList();
-        $captcha->expects($this->once())
-            ->method('getViolations')
-            ->willReturn($violations);
-
-        $listener = new CaptchaRouteListener(
-            [$captcha],
-            static::getContainer()->get(SystemConfigService::class),
-            static::getContainer()
-        );
-
-        $originalController = $event->getController();
-        $listener->validateCaptcha($event);
-
-        // Verify that the controller was changed
-        // @see CaptchaRouteListener::validateCaptcha()
-        static::assertNotSame($originalController, $event->getController());
-
-        // Verify that the new controller is callable
-        $newController = $event->getController();
-        static::assertIsCallable($newController);
-    }
-
-    /**
-     * @return array<int, AbstractCaptcha|MockObject>
-     */
-    private function getCaptchas(bool $supports, bool $isValid): array
-    {
-        $captcha = $this->getMockBuilder(AbstractCaptcha::class)->getMock();
-
-        $captcha->expects($this->once())
-            ->method('supports')
-            ->willReturn($supports);
-
-        $captcha->expects($supports ? $this->once() : $this->never())
-            ->method('isValid')
-            ->willReturn($isValid);
-
-        $captcha->expects($supports ? $this->once() : $this->never())
-            ->method('shouldBreak')
-            ->willReturn(true);
-
-        return [$captcha];
-    }
-
-    private function getControllerEventMock(): ControllerEvent
-    {
-        return new ControllerEvent(
-            $this->createMock(HttpKernelInterface::class),
-            function (): void {
-            },
-            self::getRequest($this->getRequestAttributes(true)),
-            HttpKernelInterface::MAIN_REQUEST
-        );
-    }
-
-    private static function getRequest(ParameterBag $attributes): Request
-    {
-        return new Request(attributes: $attributes->all());
-    }
-
-    private function getRequestAttributes(bool $isCheckEnabled): ParameterBag
-    {
-        $param = [
-            '_captcha' => $isCheckEnabled ? true : null,
-        ];
-
-        return new ParameterBag($isCheckEnabled ? $param : []);
     }
 }
